Removed old KOS implementation and streamlined api
