+++ mpc-core/examples/ote_random.rs
@@ -7,15 +7,11 @@
 
 use mpc_core::block::Block;
 
-pub use mpc_core::ot::extension::kos15refactor::receiver::{
-    error::ExtReceiverCoreError, BaseSend as RBaseSend, BaseSetup as RBaseSetup,
-    Initialized as RInitialized, Kos15Receiver, Setup as RSetup,
+pub use mpc_core::ot::extension::{
+    ExtReceiverCoreError, ExtSenderCoreError, Kos15Receiver, Kos15Sender, RBaseSend, RBaseSetup,
+    RInitialized, RSetup, SBaseReceive, SBaseSetup, SInitialized, SSetup,
 };
 
-pub use mpc_core::ot::extension::kos15refactor::sender::{
-    error::ExtSenderCoreError, BaseReceive as SBaseReceive, BaseSetup as SBaseSetup,
-    Initialized as SInitialized, Kos15Sender, Setup as SSetup,
-};
 pub fn main() {
     // Sender messages the receiver chooses from
     let inputs = [
