+++ mpc-core/src/ot/extension/kos15/receiver/error.rs
@@ -11,12 +11,8 @@ pub enum ExtReceiverCoreError {
     NotDerandomized,
     #[error("Tried to derandomize more OTs than setup")]
     InvalidChoiceLength,
-    #[error("Received choice length which is not a multiple of 8")]
-    ChoiceNotMultipleOfEight,
     #[error("Received payload of unexpected size")]
     InvalidPayloadSize,
-    #[error("An internal error happened")]
-    InternalError,
     #[error("Matrix Error: {0}")]
     MatrixError(#[from] MatrixError),
 }
