--- mpc-core/src/ot/extension/kos15refactor/mod.rs
@@ -1,6 +1,0 @@
-mod matrix;
-pub mod receiver;
-pub mod sender;
-pub mod utils;
-
-const BASE_COUNT: usize = 128;
